// apps/worker/ui/src/api/types.ts
export type GetListRequest = {
  offset: number;
  limit: number | null;
  selected_ids: number[];
  search_text: string | null;
  text_filter: Record<string, string[]>;
  filter: Record<string, unknown[]>;
  sort: [string, 'asc' | 'desc'] | null;
  random?: boolean;
};

export type GetListResponse<T> = {
  total: number;
  items: T[];
};

export type UpsertResponse = {
  id: number;
  fk_not_found?: Record<string, number[]> | null;
};

export type ExampleRecord = {
  id?: number;
  created_at?: string | null;
  updated_at?: string | null;
  jp_text: string;
  kr_text: string;
  context?: string | null;
  prompt?: string | null;
  negative_prompt?: string | null;
  context_embedding?: number[] | null;
  text_embedding?: number[] | null;
  prompt_embedding?: number[] | null;
  jp_words?: number[] | null;
  audios?: number[] | null;
  error_reports?: number[] | null;
};

export type SimilarityResult = {
  id: number;
  score: number;
};

export type ExampleSortSimilarRecord = ExampleRecord & {
  similar_prompt_image?: SimilarityResult | null;
  similar_context_text_example?: SimilarityResult | null;
  similar_text_context_example?: SimilarityResult | null;
};